Put jl_assume/jl_assume_aligned back in dtypes.h
