+++ src/support/dtypes.h
@@ -3,6 +3,7 @@
 #ifndef JL_DTYPES_H
 #define JL_DTYPES_H
 
+#include <assert.h>
 #include <stddef.h>
 #include <stddef.h> // double include of stddef.h fixes #3421
 #include <stdint.h>
@@ -14,7 +15,6 @@
 
 #include "platform.h"
 #include "analyzer_annotations.h"
-#include "../julia_assert.h"
 
 #if !defined(_OS_WINDOWS_)
 #include <inttypes.h>
@@ -141,6 +141,89 @@ typedef intptr_t ssize_t;
 #  define JL_ATTRIBUTE_ALIGN_PTRSIZE(x) x __attribute__ ((aligned (sizeof(void*))))
 #else
 #  define JL_ATTRIBUTE_ALIGN_PTRSIZE(x)
+#endif
+
+#if defined __has_builtin
+#  define jl_has_builtin(x) __has_builtin(x)
+#else
+#  define jl_has_builtin(x) 0
+#endif
+
+#ifdef JL_NDEBUG
+# if jl_has_builtin(__builtin_assume)
+#define jl_assume(cond)                  \
+    (__extension__({                     \
+        __typeof__(cond) cond_ = (cond); \
+        __builtin_assume(!!(cond_));     \
+        cond_;                           \
+    }))
+#elif defined(__GNUC__)
+static inline void jl_assume_(int cond)
+{
+    if (!cond) {
+        __builtin_unreachable();
+    }
+}
+#define jl_assume(cond)                  \
+    (__extension__({                     \
+        __typeof__(cond) cond_ = (cond); \
+        jl_assume_(!!(cond_));           \
+        cond_;                           \
+    }))
+#else
+#  define jl_assume(cond) (cond)
+# endif
+# if jl_has_builtin(__builtin_assume_aligned) || defined(_COMPILER_GCC_)
+#  define jl_assume_aligned(ptr, align) __builtin_assume_aligned(ptr, align)
+# elif defined(__GNUC__)
+#define jl_assume_aligned(ptr, align)               \
+    (__extension__({                                \
+        __typeof__(ptr) ptr_ = (ptr);               \
+        jl_assume(((uintptr_t)ptr_) % (align) == 0); \
+        ptr_;                                       \
+    }))
+#elif defined(__cplusplus)
+template<typename T>
+static inline T
+jl_assume_aligned(T ptr, unsigned align)
+{
+    (void)jl_assume(((uintptr_t)ptr) % align == 0);
+    return ptr;
+}
+# else
+#  define jl_assume_aligned(ptr, align) (ptr)
+# endif
+#else
+# if defined(__GNUC__)
+#define jl_assume(cond)                  \
+    (__extension__({                     \
+        __typeof__(cond) cond_ = (cond); \
+        assert(!!(cond_));               \
+        cond_;                           \
+    }))
+#define jl_assume_aligned(ptr, align)             \
+    (__extension__({                              \
+        __typeof__(ptr) ptr_ = (ptr);             \
+        assert(((uintptr_t)ptr_) % (align) == 0); \
+        ptr_;                                     \
+    }))
+# elif defined(__cplusplus)
+#define jl_assume(cond)      \
+    (([&]() {                \
+        auto cond_ = (cond); \
+        assert(!!(cond_));   \
+        return cond_;        \
+    })())
+#define jl_assume_aligned(ptr, align)             \
+    (([&]() {                                     \
+        auto ptr_ = (ptr);                        \
+        assert(((uintptr_t)ptr_) % (align) == 0); \
+        return ptr_;                              \
+    })())
+#else
+#  define jl_assume(cond) (cond)
+#  define jl_assume_aligned(ptr, align) (ptr)
+# endif
 #endif
 
 typedef int bool_t;
